Update readme.md
a). add the link of immutable.js which supports immutable data
b). As we know, `const` is not an absolute constant sometimes, such like: 

```js
const a = { a: 1 };
delete a.a;
console.log(a);
// => {}
```

Immutable data cannot be changed once created which we want and where immutable.js comes in:

```js
const { fromJS } = require('immutable');

const a = fromJS({ a: 1 });
// a => { "a": 1 }

const b = a.delete('a');
// a => { "a": 1 }
// b => {}
```

c). Are all code snippets should be rewrote by immutable.js ? Not sure about all, but it is more accurate for the code snippet of Constant.