Added reference to Crocks ADT library
crocks is a collection of popular Algebraic Data Types (ADTs) that are all the rage in functional programming. The main goal of crocks is to curate and provide not only a common interface between each type (where possible of course), but also provide all of the helper functions needed to hit the ground running.